Finalizando modulo de despesas

+++ src/pages/Expenditure/List/index.tsx
@@ -10,26 +10,42 @@ import Swal from 'sweetalert2';
 
 const ListExpenditure = (): JSX.Element => {
   const handleDelete = async (id?: string) => {
-    try{
-      const currentSession = await Auth.currentSession();
-      await api.delete(`/expenditure/${id}`, {
-        headers: {
-          'X-Cognito-ID-Token': currentSession.getIdToken().getJwtToken()
-        }
-      });
+    const currentSession = await Auth.currentSession();
 
-      Swal.fire({
-        icon: 'success',
-        title: 'Sucesso!',
-        text: 'O cliente foi excluído com sucesso!'
-      });
-    } catch (error) {
-      Swal.fire({
-        icon: 'error',
-        title: error.response.data.title,
-        text: error.response.data.text,
-      });
-    }
+    Swal.fire({
+      icon: 'question',
+      title: 'Atenção!',
+      text: 'Deseja realmente excluir a despesa selecionada?',
+      showCancelButton: true,
+      confirmButtonText: "Sim, confirmar!",
+      cancelButtonText: "Não!"
+    }).then( async (result) => {
+      if(result.isConfirmed){
+        try{
+          await api.delete(`/expenditure/${id}`, {
+            headers: {
+              'X-Cognito-ID-Token': currentSession.getIdToken().getJwtToken()
+            }
+          });
+    
+          Swal.fire({
+            icon: 'success',
+            title: 'Sucesso!',
+            text: 'A despesa foi excluída com sucesso!'
+          }).then(result => {
+            window.location.reload(false);
+          });
+        } catch(error) {
+          Swal.fire({
+            icon: 'error',
+            title: error.response.data.title,
+            text: error.response.data.text,
+          });
+        }
+      } else {
+        Swal.close();
+      }
+    });
   }
 
   return (
@@ -41,8 +57,8 @@ const ListExpenditure = (): JSX.Element => {
         formUrl="/expenditure"
         headCells={[
           {id: "description", disablePadding: true, label: "Descrição", numeric: false, type: "text" },
-          {id: "referenceDate", disablePadding: true, label: "Referência", numeric: false, type: "date" },
-          {id: "expenditureType", disablePadding: true, label: "Tipo", numeric: false, type: "text" },
+          {id: "referenceDate", disablePadding: true, label: "Referência", numeric: false, type: "dateReference" },
+          {id: "expenditureType", disablePadding: true, label: "Tipo da despesa", numeric: false, type: "text" },
           {id: "dueDate", disablePadding: true, label: "Vencimento", numeric: false, type: "date" },
           {id: "options", disablePadding: true, label: "", numeric: false }
         ]}
@@ -54,14 +70,16 @@ const ListExpenditure = (): JSX.Element => {
               link: "/expenditure",
               title: "Visualizar",
               handle: () => {},
-              icon: <VisibilityIcon />
+              icon: <VisibilityIcon />,
+              action: 'view'
             },
             {
               type: "link",
               link: "/expenditure",
               title: "Editar",
               handle: () => {},
-              icon: <CreateIcon />
+              icon: <CreateIcon />,
+              action: 'edit'
             },
             {
               type: "button",
